cleanup getQuotas
